+++ core/src/main/java/com/github/skjolber/packing/packer/util/WeightPressureCountLoadAwarePlacementUtility.java
@@ -17,8 +17,8 @@ public class WeightPressureCountLoadAwarePlacementUtility extends AbstractLoadWe
 	}
 
 	@Override
-	public long calculateSupporteeLoad(BoxStackValue sv, int minX, int minY, int minZ, int maxX, int maxY) {
-		long weight = 0;
+	public double calculateSupporteeLoad(BoxStackValue sv, int minX, int minY, int minZ, int maxX, int maxY) {
+		double weight = 0.0;
 		int z = minZ + sv.getDz();
 		int stackSize = stack.size();
 		for (int i = 0; i < stackSize; i++) {
@@ -35,12 +35,12 @@ public class WeightPressureCountLoadAwarePlacementUtility extends AbstractLoadWe
 			}
 
 			long area = LoadPlacementUtility.overlapArea(minX, minY, maxX, maxY, candidate);
-			long candidateWeight = candidate.getWeight() + candidate.getLoadWeight();
-			long effectiveWeight = (candidateWeight * area) / (area + candidate.getSupportedArea());
+			double candidateWeight = candidate.getWeight() + candidate.getLoadWeight();
+			double effectiveWeight = candidateWeight * area / (area + candidate.getSupportedArea());
 
 			if (sv.isMaxLoadPressure()) {
 				if (Box.calculatePressure(area, effectiveWeight) > sv.getMaxLoadPressure()) {
-					return -1;
+					return -1.0;
 				}
 			}
 
@@ -49,7 +49,7 @@ public class WeightPressureCountLoadAwarePlacementUtility extends AbstractLoadWe
 		}
 
 		if (sv.isMaxLoadWeight() && weight > sv.getMaxLoadWeight()) {
-			return -1;
+			return -1.0;
 		}
 		return weight + sv.getBox().getWeight();
 	}
